Linear search

+++ 0x1E-search_algorithms/0-linear.c
@@ -13,7 +13,7 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	unsigned int i = 0;
+	int i = 0;
 
 	if (array == NULL)
 	{
@@ -24,7 +24,7 @@ int linear_search(int *array, size_t size, int value)
 		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 		{
-			return ((int)i);
+			return (i);
 		}
 	}
 	return (-1);
